Add ability to define form options as functions
