+++ src/apps/components/form/controllers.js
@@ -1,28 +1,7 @@
-const metadata = require('../../../lib/metadata')
-const { transformObjectToOption, transformStringToOption } = require('../../transformers')
 const { macros } = require('./constants')
-const { buildFormWithOptions } = require('../../builders')
-
-const foreignOtherCompanyOptions = [
-  'Charity',
-  'Company',
-  'Government dept',
-  'Intermediary',
-  'Limited partnership',
-  'Partnership',
-  'Sole trader',
-]
 
 function renderFormElements (req, res) {
   res.locals.macros = res.locals.macros || macros
-
-  res.locals.macros.standardForm = buildFormWithOptions(res.locals.macros.standardForm, {
-    country: metadata.countryOptions.map(transformObjectToOption),
-    sector: metadata.sectorOptions.map(transformObjectToOption),
-    strategicDrivers: metadata.strategicDriverOptions.map(transformObjectToOption),
-    averageSalary: metadata.salaryRangeOptions.map(transformObjectToOption),
-    foreignOtherCompany: foreignOtherCompanyOptions.map(transformObjectToOption),
-  })
 
   return res
     .breadcrumb('Form elements')
@@ -30,15 +9,6 @@ function renderFormElements (req, res) {
       macros: res.locals.macros || macros,
       entitySearch: Object.assign({}, res.locals.entitySearch, {
         searchTerm: req.query.term,
-      }),
-      form: Object.assign({}, res.locals.form, {
-        options: {
-          countries: metadata.countryOptions.map(transformObjectToOption),
-          averageSalaryRange: metadata.salaryRangeOptions.map(transformObjectToOption),
-          strategicDrivers: metadata.strategicDriverOptions.map(transformObjectToOption),
-          sectors: metadata.sectorOptions.map(transformObjectToOption),
-          foreignOtherCompany: foreignOtherCompanyOptions.map(transformStringToOption),
-        },
       }),
     })
 }
